Numbers_Suffix/requirements.txt
datrie==0.8.2
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
